Enforce that only one break is added on a #get call. This seems like a good discipline that we already almost followed, except for add, sequence and generator based sources. Add and sequence are fixed, the merging of tracks that they are intended to perform now come with the deletion of breaks. Generators were quickly fixed but they still need a more general cleanup.
An extra fix for add (only for audio) is that it blanks the frame if needed
before starting to add data over it -- this happens if one source has more
data to put than the others.

This ref #204 and fixes the minimal example I gave. But it only does so
by avoiding meaningless breaks -- those left by a merging sequence in cross.
The loss of metadata is likely to remain.

--HG--
extra : convert_revision : svn%3Aaec24677-d710-0410-a355-ac75e2bdf181/trunk%406232