Allow for garbage collection of inner clocks
The inner clocks of operators such as cross(), cue_cut(), etc. were allocated
but could never be claimed by the GC. After the source is destroyed, its clock
was still referenced by the outer clock (typically the wallclock, which is not
collected). Fixed this by adding the possibility to detach a subclock from a
concrete clocks as well as clock variables, and using it in all operators that
create inner clocks.