Cleanup use of polymorphic variants in Harbor
A few assert false were totally useless. Others were caused by a partial
implementation of string_of_protocol (partial at least wrt. the type of the
protocol which was being converted to a string). It seems to me that my
changes are backward compatible but I badly need a review here, I have no
idea of the intended invariants; it could very well be that protocol/verb
types should be constrained in several places instead of trying to cover
all the cases as I did.