Compute kinds by unification in order to get rid of types during evaluation. (#1202)
Remove all types at execution.

The typing phase is left unchanged but after that we don't mention types anymore. The reasons why it is a good thing are that

- this speeds up execution (execution times x3 higher because we don't have to maintain types - which is quite costly, for instance each time we use a variable we have to instantiate its type which requires going through the whole type and copying it, each time we apply a function we have to perform unification, etc.)
- the kernel is simpler: for instance generalization had to consider all types occurring in all terms, whereas we can simply consider the type of the current term now, as usual
- the kernel of the language is smaller and thus safer and more modular
- as a particular cas of the above point, I needed this in for record (#1197) which would be hell to implement otherwise.
The only place where types were needed are in order to compute the number of (audio/video/midi) channels to produce for operators. This is now inferred by unification when creating the operators (I use the same mechanism than for clocks). The PR also simplifies much the typing of channels: they roughly all have the same type whereas we had something mildly different to type frame kinds and source kinds, etc. We could go even further, but I didn't want to ruin the readability of this PR.

The downside. We essentially do the work twice: we check that the typing of channels is correct at typing time, and once again at runtime. In my opinion, this could be entirely removed from typing, since unification is performed right at the beginning of the execution of the program which is roughly the same as typing phase. The only difference is for dynamically created sources, but those aren't used much (for now). And there is no reason to incorporate the number of channels in typing, but not say size of images, samplerate, etc. Let's keep that at runtime. I did not remove it for now because our users are "used" to the current types source('a,'b,'c) (to be honest, I am not even sure of that: many of them would simply prefer source...)