Enhance typing of lists.
We simplify the typing of lists.

If the simple method does not work, we try to remove methods in order not to
uselessly force methods. Before that the following was not working:

```
queue = request.queue()
radio = fallback([queue, playlist])
```

because `queue` has a method `push` and `playlist` does not have it.