Pass metadata to ffmpeg filters.
This is a bit tricky because our metadata are global while ffmpeg's are per-stream. We make the assumption that metadata matter more for audio data where it usually carries song title/artist/.. so we pass metadata back and forth to audio streams only by default. A `pass_metadata` parameter is added to `ffmpeg.filter.*.{input,output}` to give the user the choice to change this default.