Increasing types (#1905)
Implement a variant of subtyping where substituted type variables carry variance information and can subsequently be changed.

The typical use case is
```
if ... then 1 else null() end
```
Here the `if` takes two branches of type `'a` and returns `'a`. When we look at the first branch, we determine that `'a` must be `int` and therefore we raise a type error when we look at the second branch because we don't have `'b? <: 'a`. With this proposal, we record, when we substitute `'a` by `int`, that any greater type would have been ok too, and therefore allow ourselves changing, `int` to `int?` (which is ok since `int <: int?`).

It should also solve the issue of lists with heterogeneous (but compatible) elements (e.g. the elements have some methods in common but not all).

In more details, let me try to explain better why you need the variance information. Basically, a `'a` variable can be replaced by anything, so we have all freedom there. When you perform a unification `int <: 'a`, we replace `'a` by `int` but we keep the information that anything greater than `int` would also do (such as `int?`). However, when the comparison is in the other direction `'a <: int`, we also replace `'a` by `int`, but this time, we cannot increase `'a` after since it might not be smaller than `int` afterward. In this case, we mark the variance as `Invariant` so that we know that we don't have the right to change it. We could try to mark it as `Contravariant` and use this information, but I don't want to think too hard for now (e.g. what should we do if we have a covariant link to a contravariant link to something?...). So either we allow increasing, or we don't.

Things with constraints are marked as invariant, although they could sometimes be increasing, because after increasing we should check the constraint again, and I am too lazy to do that for now ;)

Concerning the tests, there are already quite a few of them:
```
  _ = if true then {a = 4, b = 5} else {a = 4} end
  _ = if true then {a = 4} else {a = 4, b = 5} end
```
```
 def f() = if true then null() else 1 end end
  def f() = if true then 1 else null() end end
  _ = [null(), 3]
  _ = [3, null()]
```
etc. And with our "large" codebase, it's already quite tricky to have stdlib passing ;)