Do not reset the generator mode to `Undefined in srt input on each reconnection. * Generator is created as `Undefined mode * Mode is set via the call to mk_buffer when filling from the frame.
There might have been a time where frame content could change over time but this is no longer the case so setting it once should be enough.

This is in particular an issue when dealing with reconnection. It is possible for a disconnection to occur while the audio and video content in the buffer are not of the same length. This makes the generator crash when changing mode.

Fixes: #1950, #1186