Make the decoder follow the dimensions indicated by the format. (#2212)
Video dimensions (width and height) can now be specified per stream in the type and are then used instead of the default ones. For instance, you can now write

```
s = (single("file.mp4") : source(video(width=300,height=200)))
```

in order to force the decoding of a file to be performed at the 300×200 resolution.