Correct list.shuffle.
Use Fisher-Yates to implement `list.shuffle` instead of `list.sort`
with random comparison function: `List.sort` is expecting a total order,
which is not what we provide here. The correct implementation could be
found before as `list.randomize`, I have changed its name to
`list.shuffle` so that we have only one implementation (and presumably
nobody was using `list.randomize`). This was used to shuffle playlists
and should therefore fix related bugs.

Fixes #2057, #2500.