Autocue - Introduced sustained ending support
This is an improvement of autocue to optimize crossfade and cue out points in case of sustained endings. This new algorithm is analysing the loudness behavior after the initially calculated crossfade and cue out point. If it's still a part of the song that is likely to be of relevance, the crossfade is recalculated based on that data.

The relevance is determined on the basis of the loudness reduction immediately after the original crossfade point and a possible "plateau" (i.e. no continuous fade-out).