Fix telnet functions for request.queue
There are issues with the telnet functions because they do not use the
overwritten methods of the request.queue implementation. For example,
`flush_and_skip` does not actually clear the entire queue since it does
not use the `set_queue` defined below it which clears the internal
s.queue() and modifies the backing queue. Since it does not set the
backing queue to empty, any requests that were in the backing queue
are left over.